Product Serivce Impl tested

+++ src/main/java/entity/Boots.java
@@ -4,7 +4,7 @@ public class Boots extends Product {
     private int size;
     private boolean isNatutalSkin;
 
-    public Boots(int id, String productName, double price, double weight, int productCount, int size, boolean isNatutalSkin) {
+    public Boots(Long id, String productName, double price, double weight, int productCount, int size, boolean isNatutalSkin) {
         super(id, productName, price, weight, productCount);
         this.size = size;
         this.isNatutalSkin = isNatutalSkin;
