+++ src/main/java/entity/Cloth.java
@@ -4,7 +4,7 @@ public class Cloth extends Product {
     private String size;
     private String material;
 
-    public Cloth(int id, String productName, double price, double weight, int productCount, String size, String material) {
+    public Cloth(Long id, String productName, double price, double weight, int productCount, String size, String material) {
         super(id, productName, price, weight, productCount);
         this.size = size;
         this.material = material;
